requirements.txt
# Core dependencies
pyyaml>=6.0
boto3>=1.26.0
gitpython>=3.1.0
click>=8.0.0
colorama>=0.4.0
rich>=13.0.0

# Configuration and validation
jsonschema>=4.0.0

# Utilities
requests>=2.28.0